refactor: centralize zip extraction logic in `File.unzipTo`
- Add `unzipTo` extension function to `File` in `FileUtil.kt` to handle zip file extraction
- Update `FileAction.kt` and `XedPackage.kt` to use the new utility instead of manual inline zip extraction logic
- Remove unused `java.util.zip.ZipFile` imports from refactored files
