+++ core/main/src/main/java/com/rk/filetree/FileAction.kt
@@ -12,6 +12,7 @@ import com.rk.activities.main.MainActivity
 import com.rk.drawer.DrawerViewModel
 import com.rk.file.FileObject
 import com.rk.file.FileOperations
+import com.rk.file.unzipTo
 import com.rk.icons.CreateNewFile
 import com.rk.icons.CreateNewFolder
 import com.rk.icons.Icon
@@ -26,7 +27,6 @@ import com.rk.utils.toast
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import java.io.File
-import java.util.zip.ZipFile
 
 data class FileActionContext(
     val file: FileObject,
@@ -291,22 +291,7 @@ object UnzipAction : FileAction() {
 
             runCatching {
                 context.viewModel.withFileOperation {
-                    // TODO: Write extract method for file
-                    ZipFile(zipFile).use { zip ->
-                        zip.entries().asSequence().forEach { entry ->
-                            val entryFile = File(targetDir, entry.name)
-                            if (entry.isDirectory) {
-                                entryFile.mkdirs()
-                            } else {
-                                entryFile.parentFile?.mkdirs()
-                                zip.getInputStream(entry).use { input ->
-                                    entryFile.outputStream().use { output ->
-                                        input.copyTo(output)
-                                    }
-                                }
-                            }
-                        }
-                    }
+                    zipFile.unzipTo(targetDir)
                 }
             }
                 .onSuccess {
